fix: don't assume the step of a for-loop is a generation-time constant
This is a bit of a happy accident here. I thought we were explicitly requiring
the step of a for loop to be a generation-time constant, as CMurphi does.
However, it turns out we were actually constant folding the step during
validation, and as a side effect throwing an error during constant folding when
we encountering something non-constant.

This commit fixes validation to only attempt constant folding of the step when
it is actually foldable. We still need some generated validation checks for
loops that have dynamic bounds and/or steps.