avoid assertion failure when the user sets a small type for --value-type
When building in debug mode, the user could trigger an assertion failure by
setting a value type that was too small to contain values in the model. Note
that this would *only* occur when the user specifically overrode --value-type
with a narrower type than the one that would have been selected automatically.

Github: closes #166 "users can force verifier assertion failure"