fix: accept type-compatible return values though not type-equal
We were rejecting functions that returned values of range type when the type of
the returned expression was not directly a range. This could happen when the
type of the returned expression was a TypeExprID that then resolved to a range.
We now accept such constructions.

Github: closes #169 "returning type-compatible but differing range from a
  function is rejected"