implement non-const iteration on Nodes
We provide an API supporting range-based for loops and legacy iteration over an
AST or a subtree rooted at a particular node. The purpose of this is to abstract
patterns like indexing and validation into a common implementation. This needs a
const counterpart, but then it should be ready to near-entirely replace the
functionality of indexing, validation, and friends. This should then be a
suitable base to build Node::all, Node::any, Node::none, etc on top of.

Github: related to #152 "traversal abort"