murphi2c: emit all switches as if statements
There are some subtle cases where switch cases that seem constant are
technically not according to the C standard. The scenario induced by
tests/const-enum.m is something like:

  typedef enum { A, B } T;
  const T x = A;

  switch (y) {
    case x: ...
  }

Clang seems fine with this, but GCC does not like it. To work around this we now
simply transform all switch statements.