use binutils 2.26 on Ubuntu 14.04 when testing GCC 7-9
As per the .travis.yml comment, we need to use binutils ≥ 2.26 if we want to
generate -march=native code. The test suite does not currently do this, but we
would like to introduce some tests for rumur-run which does pass these flags.

Github: related to #167 "switch all Python dependencies to Python 3 in
  packaging/debian"
Github: related to #170 "optimise testing"