reduce the valgrind/not-valgrind tests to a single option
Instead of running each model and ast-dump test case under Valgrind and also not
under Valgrind, we now do only one of these: Valgrind if Valgrind is available
and not otherwise. There's no point running the non-Valgrind test if we've
already run the Valgrind test and the test case skipping that happens when we
don't have Valgrind simply slows the test suite.

Github: related to #170 "optimise testing"