guard Ubuntu 14.04 GCC 7+ check to only take effect in Travis CI
This check was introduced to suppress false failures in some of the Travis CI
configurations. However, it causes problems during Debian packaging because
pbuilder tries to run lsb_release which is not accessible within the pdebuild
chroot. To avoid this, we prevent reaching the lsb_release check unless we are
running in Travis.

Github: related to #167 "switch all Python dependencies to Python 3 in
  packaging/debian"