fix: link against libatomic when on a platform without dword CAS
Packaging v2019.12.22 for Debian revealed that the test suite fails on a number
of platforms [0] due to no double-word compare-and-swap available in the ISA. To
avoid this we now link against libatomic on such platforms. This will not
produce lock-free code and is somewhat undesirable, but it seems preferable to
not supporting such platforms at all.

  [0]: https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=mips64el&ver=2019.12.22-1&stamp=1577335749&raw=0,
       https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=mipsel&ver=2019.12.22-1&stamp=1577341073&raw=0,
       https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=s390x&ver=2019.12.22-1&stamp=1577329999&raw=0,
       https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=powerpc&ver=2019.12.22-1&stamp=1577331185&raw=0,
       https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=ppc64&ver=2019.12.22-1&stamp=1577331096&raw=0,
       https://buildd.debian.org/status/fetch.php?pkg=rumur&arch=sparc64&ver=2019.12.22-1&stamp=1577330648&raw=0