deprecate Node::operator== and vector_eq
The existence of a public operator== was confusing to users because it did not
implement the exact intuitive notion of equality they expected. Moreover, it is
not clear how to implement such a thing within an AST. E.g. are two identical
entities that come from different files equal? To avoid confusion, we now
discourage using them altogether.

Github: closes #155 "remove equality comparators"