move some handle reading functionality earlier
The aim here is to get to a point where this functionality is usable for reading
a packed version of state.bound. Note that the new read_raw function supports
reading anything up to a uint64_t regardless of what our value type is.

Github: related to #158 "pack state struct"