avoid emitting state.liveness if there are no liveness properties
This was previously emitted as a 0-element array in this scenario. While this
did not enlarge the state struct, it forced its alignment to be >= 8 bytes. By
removing it entirely we can reduce state struct alignment to 1 byte on x86-64
Linux and 2 bytes on other platforms. The overall aim here is to enable the
state allocator to more efficiently allocate states in memory.

Github: related to #158 "pack state struct"