use a specific value for the tombstone instead of a bit
We never actually need to recover the original pointer from a tombstone. We can
take advantage of this to just use a sentinel value (-1) to represent a
tombstone. This should speed up any testing against or setting of tombstones,
but it also removes alignment requirements we have on the state struct.