xxd: support UTF-8, binary data
When using UTF-8 characters in an input to xxd.py, multi-byte characters would
result in invalid code generation. We now support any arbitrary bytes in the
input file. Though none of the resources we translate with xxd.py currently
contain multi-byte characters, this allows us to use them in future.