fix incorrect condition for constructing AND mask in write_raw()
When a handle's aligned width worked out to be 64 bits but its actual width and
offset did not fully extend to 64 bits, the upper bits of the word would be
incorrectly zeroed. Though this could have affected actual models, it is mostly
a latent bug as no known models used handles this wide with these
characteristics. This bug was found during construction of the write_raw()
proof.