murphi2c: fix: index the model
Despite not indexing the model, we were using unique identifiers to track
typedefed enums. The result of this was that we only ever tracked a single enum
(the last one typedefed) and this matched every encountered eum constant. The
effect of this would be that, in a model with multiple enum types, any enum
constants might end up with the incorrect type in murphi2c output. This caused a
-Wenum-conversion warning, but otherwise was not obvious to the user.