reorder fields to optimise bit offsets during checking
Because the specific ordering of fields within a record or within a model is not
relied on, we now reorder these to try to optimise for performance of runtime
reads and writes during verification. Aligned hardware accesses are typically
faster than unaligned accesses, and we can avoid doing bit shifting in software
if we have a natural alignment. We try to approach the optimal here, but it is
essentially best effort.

Github: closes #147 "optimise record and state field ordering"