another attempt at solving Python 3.8 macOS issues
It's unclear what the breaking change is in Python 3.8 now, but the macOS builds
were still failing, this time claiming that config was None. This time we try to
work around it by not relying on any complex globals and instead storing the
(duplicate) config within the test cases themselves.