murphi2xml: re-index the model prior to symbol resolution
Technically you can get away without re-indexing the model and still running
symbol resolution. However, the unique identifiers will be all screwed up after
this and certain analyses on the AST won't work (e.g. checking whether a
function is pure). The XML generator doesn't rely on any of this functionality
but it seems safer to re-index anyway in case we add functionality in future
that does depend on this.