implement new synthetic AmbiguousAmp node
This new AST node is designed to represent a 'x & y' expression before we have
decided whether it is a logical AND or a bitwise AND. The intention here is to
move towards a situation where we can overload this operator to mean either.
This should be achievable as we have a stronger type system than C, so we should
be able to come up with precise types for both operands.

Github: related to #37 "extended operator support"