fix: when reordering record fields, also look inside ExprID referents
We were not descending into the referents of ExprIDs when optimising record
field ordering, which could have the effect of reordering a record's fields in
one place but not in others. This is sort of a problem of our own making, in
that it only exists because we copy types during symbol resolution rather than
taking a weak reference to them. I still think taking copies is the way to go
and results in a simpler AST, but it's unfortunate that we open ourselves up to
problems like this.

Github: related to #193 "State corruption with scalarset?"