rename 'schedule_*' to 'stack_*' in exhaustive canonicalisation
This code is implementing Heap's algorithm for generating all permutations of a
given range. The parameter that we previously called "schedule" is what is known
as "c" or the "stack" in the pedagogical description of the algorithm. Renaming
it helps avoid confusion, now that we're introducing something else called
"schedule" that is a concept of our own invention.

Github: related to #194 "clearer CEX output with scalarsets"