rephrase a generate_print call to Generator()
This has no immediate relevance, but we're about to make the `prefix` parameter
to Generator a more complex type than std::string. This particular call site
will need to take advantage of this to pass through richer information than a
simple string that it can pass to generate_print().

Github: related to #194 "clearer CEX output with scalarsets"