generate loops to print range- and scalarset-indexed arrays
Previously such arrays were unrolled at code generation time. This was
suboptimal for large arrays (e.g. an array with an index type of `0 .. 10000`
would generate a huge amount of code to print it). However, the main motivation
for this change is that we would like to print scalarsets using more friendly
identifiers. This requires us to emit printing as a loop, because the order of
the elements is not known at code generation time.

Github: related to #194 "clearer CEX output with scalarsets"