retrieve and use the *previous* schedule when diffing scalarset-indexed arrays
This is the final change that brings diff counterexample traces (the default)
inline with the user's expectations. Traces should now correctly suppress
printing of elements that have not changed and preserve elements that have
changed. The effect can be seen on tests/scalarset-cex.m.

Github: related to #194 "clearer CEX output with scalarsets"