fix: aliases of aliases generating incorrect code
Defining an alias in an aliasrule whose target was a preceding alias in the same
rule would cause incorrect code to be generated. Code for the aliases would be
emitted in the reverse order, causing the later alias to reference variables
that had not yet been defined.

Github: fixes #199 "aliases of aliases within aliasrules do not work"