fix message truncation when printing errors in XML mode
The construction of the output message in error() had an off-by-one bug when
running in XML mode (`--output-format machine-readable`). This would result in
error messages having their last character dropped. Interestingly I only noticed
this problem when I replicated this logic somewhere else for a different purpose
and noticed the last character of my messages missing.