fix print code for values in murphi2c
This seems to have been broken more or less forever, but apparently C compiler's
accepted this as a non-prototyped function. This was exposed as an error by a
recent transition to Apple clang version 12.0.0 which defaults to C99, requiring
stricter rules around pre-prototyping functions. We should probably set a C
standard explicitly during compilation in testing.