use thread fences instead of relaxed atomics for managing seen_count
Switching to relaxed atomics actually resulted in a performance *degradation*.
This is an alternate attempt at performance optimisation where we remove the use
of atomics in the hot path of set_insert(). A side effect of this change is that
we lose accuracy in reported set occupancy. Both the progress output and the
decision on when to expand the seen set are now approximate. However, the final
summary of seen states is still precise.

Github: related to #117 "thread-local set size"