revert the CI move to afl++
We moved from building the latest version of AFL from source, to using the afl++
package in Ubuntu Focal. However, this package lately does not seem to work.
Specifically it is unable to find libgmp when used as a compiler in combination
with CMake. Reverting to the old AFL setup appears to fix this.

This reverts 4944427734628cf913e8d5eeb54d897033f9eb59.