move macOS CI testing to Cirrus
CI for macOS on Travis is dead [0], so we need to migrate elsewhere. This change
moves the Homebrew jobs; Macports still to come. In doing so, this drops support
for anything prior to XCode 11.3 and extends testing to XCode 12.1. Note that we
also drop the use of Ninja on FreeBSD because it was another dependency to
install and complicate the process.

 [0]: https://blog.travis-ci.com/2020-11-02-travis-ci-new-billing