fix macro expansion applied to struct field names in murphi2c output
When code inside an aliasstmt block or an alias rule referenced a record field
with the same name as the alias, the code generated by murphi2c would
incorrectly emit the alias as a macro that matched not only references to this
alias but also the field name. We solve this by using object macros for aliases
now.

This fixes part of #207. There is a similar problem with references to fields
inside a function with a var parameter of the same name. This will require a
separate solution.

Github: related to #207 "aliases that match field names cause malformed murphi2c
  code"