fix: avoid platform-specific isspace differences
Currently debugging a curious problem wherein the murphi2c test of
tests/comment-escape.m passes on FreeBSD and macOS but fails on Linux. My
suspicion is that Glibc's isspace is returning something non-zero but with the
LSB unset for isspace(' ') which, when ANDed with the promoted value of `true`,
results in 0. I don't know if this hypothesis is correct or not, but it seems
wise to avoid these sort of bitwise operations on pseudo boolean values anyway.