accept non-egregious clang-format changes
Time to stop caring about formatting. The one thing we unfortunately cannot
abide is moving the #include of location.hh above cstddef. Flex on some
platforms generates a location.hh that does not include stddef.h but uses
size_t, so it causes a compilation failure if we rearrange the headers.