fix: ensure VALUE_C is strictly a value_t
Similarly to the previous commit, when using a fast type (e.g. int_fast8_t) the
backing type may not be the precise fixed width type (e.g int8_t). This can
result in VALUE_C(x) not actually producing a value of value_t type. There is no
fast type equivalent to e.g. INT8_C -- and if there was, we would not want to
use it as it might be overly permissive -- so we address this by simply casting
the VALUE_C result.