more explicit linkage on API symbols
The symbols that are intended to be part of the public librumur API are now
qualified with visibility("default") when compiling the library and then
visibility("hidden") when including the public rumur.h header. The exception are
symbols that need Run-Time Type Information (RTTI), exceptions that are thrown
and classes that may need to be dynamically disambiguated. These need to always
have default visibility to work across library boundaries.