limit CI on ARM64, Power, s390
There is not enough time within Travis CI limits (50 min) to run the entire test
suite. We could break these jobs into multiple jobs, as is done for the main
x86-64 tests, but it does not seem worth it when the only relevant thing that is
platform-dependent is the generated verifier. So this change limits CI on these
platforms to a subset of tests that seem most relevant.