murphi2uclid: handle var parameters in a function call on RHS of assignment
This may need some tweaking, but I think this is roughly the right direction of
how to do this. It is going to regularly result in Uclid5 code that does not
compile, but I am not sure what alternative exists. There is a flexibility in
Murphi calls that does not seem mappable to Uclid5 in the general case.