murphi2uclid: handle var parameters in calls with ignored results
This deals with an obscure case where the model calls a function that has var
(mutable) parameters, but the call site is ignoring the return value. This is
demonstrated in e.g. tests/reference-function-parameter3.m. There are valid use
cases for doing this and it is not even unusual, except in the context of
murphi2uclid. The specifics of how we do the translation make this an
interesting case to deal with.