Revert "make a libutil for common functionality"
This reverts commit 7637790e643cdf11a45bf8a13f13b37da64de718. It seems I
misunderstood how to use an object library and you are not supposed to try to
link against it. Everything works out on macOS, but fails on older CMake on
Linux. It is possible to make this work, but we would still be left messing with
include paths. It seems simpler to just revert to the old configuration which,
though it builds some files multiple times, is simpler to understand and more
portable.