backport make-version to Python < 3.6
This is a bit silly and frustrating, but we are just on the edge of losing GCC 4
series support because 4.7-4.9 seem only installable in environments where
Python 3.6 is not easy to obtain. It seems a bit awkward to require Python 3.6
just for a few build scripts when everything else is backwards compatible to
that era.