add Cirrus testing on GCC 4.9
This started as an attempt to add testing on GCC 4.7. However, the GCC 4.7
Docker image is based on Debian Wheezy which is well out of support. Just
getting functioning apt repositories is challenging. Once that is bootstrapped,
it becomes apparent that the available versions of Bison, CMake and Python are
too old. We have already bypassed CMake (downloading a static binary from
Kitware) and can probably bypass Bison, but Python is on 3.2. Compiling Python
3.4 in CI is not reasonable and further backporting the Rumur Python scripts to
work with Python 3.2 is a bridge too far. The GCC 4.8 image is similarly
impoverished.

So I think, out of practicality, GCC 4.9 has to become the new supported
baseline. 4.9.0 was released in 2014, so hopefully it is not too much of a
burden for users. The GCC 4.9 image is on Debian Jessie which is functional, for
now.