drop fuzz testing in CI
This job has been broken for some time – even prior to the Travis shutdown – due
to environment problems. When working, it also had not surfaced any bugs in a
long time. While fuzzing support is there and it works, AFL is no longer a state
of the art fuzzer and fuzzing from scratch is no longer particularly effective.
A better approach to this now would be to use a more advanced fuzzer and do
continuous (or at least resuming) fuzzing.

This completes the migration away from Travis CI.