CI: remove GCC 4.9 testing
It no longer seems feasible to create a working GCC 4.9 CI environment.

Any GCC Docker image prior to 7.1 is built on Debian Jesse that is old enough
its package repositories no longer resolve. This prevents installing Rumur
dependencies, in particular Python 3 that is too onerous to build from source.

My initial attempt to get around this was to move to a later GCC image that is
built on a more recent Debian and then install GCC 4.9 into that. GCC 4.9 is not
packaged in the repositories of more recent Debian, so we need to build it from
source. Attempting this on any image beyond GCC 7 runs into the problem that GCC
4.9 can only be built with GCC ≤ 7.

So I try to use the GCC 7.1 image. For this, we need the GCC 4.9 dependencies,
libgmp, libmpc, and libmpfr installed. Building and installing GCC 4.9 works
(after disabling multilib and the Java back end). But GCC did an ABI break in
GCC 6. So GCC 4.9 cannot build Rumur against a libgmp that was built with the
newer ABI. So we need a libgmp with the newer ABI installed in order to build
GCC 4.9 with GCC 7.1 and then we need a libgmp with the older ABI in order to
build Rumur with GCC 4.9. This is doable but seems one too many hurdles to jump
through.